Moved openssl and native into their own files
