+++ libsignal-protocol/src/crypto/mod.rs
@@ -1,5 +1,13 @@
+#[cfg(feature = "crypto-native")]
+mod native;
+#[cfg(feature = "crypto-native")]
+pub use self::native::DefaultCrypto;
+
 #[cfg(feature = "crypto-openssl")]
-use std::sync::Mutex;
+mod openssl;
+#[cfg(feature = "crypto-openssl")]
+pub use self::openssl::OpenSSLCrypto;
+
 use std::{
     convert::TryFrom,
     os::raw::{c_int, c_void},
@@ -69,227 +77,6 @@ pub trait Crypto {
         iv: &[u8],
         data: &[u8],
     ) -> Result<Vec<u8>, InternalError>;
-}
-
-#[cfg(feature = "crypto-native")]
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct DefaultCrypto;
-
-#[cfg(feature = "crypto-openssl")]
-pub struct OpenSSLCrypto {
-    hmac_ctx: Mutex<Option<openssl::hash::Hasher>>,
-    sha512_ctx: Mutex<Option<openssl::hash::Hasher>>,
-}
-
-#[cfg(feature = "crypto-native")]
-impl Crypto for DefaultCrypto {
-    fn fill_random(&self, buffer: &mut [u8]) -> Result<(), InternalError> {
-        use rand::RngCore;
-        let mut rng = rand::thread_rng();
-        rng.fill_bytes(buffer);
-        Ok(())
-    }
-
-    fn hmac_sha256_init(&self, _key: &[u8]) -> Result<(), InternalError> {
-        unimplemented!()
-    }
-
-    fn hmac_sha256_update(&self, _data: &[u8]) -> Result<(), InternalError> {
-        unimplemented!()
-    }
-
-    fn hmac_sha256_final(&self) -> Result<Vec<u8>, InternalError> {
-        unimplemented!()
-    }
-
-    fn sha512_digest_init(&self) -> Result<(), InternalError> {
-        unimplemented!()
-    }
-
-    fn sha512_digest_update(&self, _data: &[u8]) -> Result<(), InternalError> {
-        unimplemented!()
-    }
-
-    fn sha512_digest_final(&self) -> Result<Vec<u8>, InternalError> {
-        unimplemented!()
-    }
-
-    fn encrypt(
-        &self,
-        _cipher: SignalCipherType,
-        _key: &[u8],
-        _iv: &[u8],
-        _data: &[u8],
-    ) -> Result<Vec<u8>, InternalError> {
-        unimplemented!()
-    }
-
-    fn decrypt(
-        &self,
-        _cipher: SignalCipherType,
-        _key: &[u8],
-        _iv: &[u8],
-        _data: &[u8],
-    ) -> Result<Vec<u8>, InternalError> {
-        unimplemented!()
-    }
-}
-
-#[cfg(feature = "crypto-native")]
-impl Default for DefaultCrypto {
-    fn default() -> Self { Self }
-}
-
-#[cfg(feature = "crypto-openssl")]
-impl Default for OpenSSLCrypto {
-    fn default() -> Self {
-        Self {
-            hmac_ctx: Mutex::new(None),
-            sha512_ctx: Mutex::new(None),
-        }
-    }
-}
-
-#[cfg(feature = "crypto-openssl")]
-impl OpenSSLCrypto {
-    fn crypter(
-        &self,
-        mode: openssl::symm::Mode,
-        cipher: SignalCipherType,
-        key: &[u8],
-        iv: &[u8],
-        data: &[u8],
-    ) -> Result<Vec<u8>, InternalError> {
-        use self::SignalCipherType::*;
-        use openssl::symm::{Cipher, Crypter};
-        let signal_cipher_type = match (cipher, key.len()) {
-            (AesCtrNoPadding, 16) => Cipher::aes_128_ctr(),
-            (AesCtrNoPadding, 24) => {
-                let nid = openssl::nid::Nid::AES_192_CTR;
-                Cipher::from_nid(nid)
-                    .expect("OpenSSL should have AES_192_CTR !!")
-            },
-            (AesCtrNoPadding, 32) => Cipher::aes_256_ctr(),
-            (AesCbcPkcs5, 16) => Cipher::aes_128_cbc(),
-            (AesCbcPkcs5, 24) => {
-                let nid = openssl::nid::Nid::AES_192_CBC;
-                Cipher::from_nid(nid)
-                    .expect("OpenSSL should have AES_192_CBC !!")
-            },
-            (AesCbcPkcs5, 32) => Cipher::aes_256_cbc(),
-            _ => unreachable!(),
-        };
-        let block_size = signal_cipher_type.block_size();
-        let mut result = Vec::with_capacity(data.len() + block_size);
-        let mut crypter = Crypter::new(signal_cipher_type, mode, key, Some(iv))
-            .map_err(|_e| InternalError::Unknown)?;
-        crypter
-            .update(data, &mut result)
-            .map_err(|_e| InternalError::Unknown)?;
-        crypter
-            .finalize(&mut result)
-            .map_err(|_e| InternalError::Unknown)?;
-        Ok(result)
-    }
-}
-
-#[cfg(feature = "crypto-openssl")]
-impl Crypto for OpenSSLCrypto {
-    fn fill_random(&self, buffer: &mut [u8]) -> Result<(), InternalError> {
-        openssl::rand::rand_bytes(buffer).map_err(|_e| InternalError::Unknown)
-    }
-
-    fn hmac_sha256_init(&self, _key: &[u8]) -> Result<(), InternalError> {
-        let nid = openssl::nid::Nid::HMACWITHSHA256;
-        let ty = openssl::hash::MessageDigest::from_nid(nid)
-            .ok_or_else(|| InternalError::Unknown)?;
-        let ctx = openssl::hash::Hasher::new(ty)
-            .map_err(|_e| InternalError::Unknown)?;
-        let mut guard =
-            self.hmac_ctx.lock().map_err(|_e| InternalError::Unknown)?;
-        *guard = Some(ctx);
-        Ok(())
-    }
-
-    fn hmac_sha256_update(&self, data: &[u8]) -> Result<(), InternalError> {
-        let mut guard =
-            self.hmac_ctx.lock().map_err(|_e| InternalError::Unknown)?;
-        if let Some(ref mut ctx) = *guard {
-            ctx.update(data).map_err(|_e| InternalError::Unknown)?;
-        }
-        Ok(())
-    }
-
-    fn hmac_sha256_final(&self) -> Result<Vec<u8>, InternalError> {
-        let mut guard =
-            self.hmac_ctx.lock().map_err(|_e| InternalError::Unknown)?;
-        if let Some(ref mut ctx) = *guard {
-            ctx.finish()
-                .map(|buf| buf.as_ref().to_owned())
-                .map_err(|_e| InternalError::Unknown)
-        } else {
-            Err(InternalError::Unknown)
-        }
-    }
-
-    fn sha512_digest_init(&self) -> Result<(), InternalError> {
-        let ty = openssl::hash::MessageDigest::sha512();
-        let ctx = openssl::hash::Hasher::new(ty)
-            .map_err(|_e| InternalError::Unknown)?;
-        let mut guard = self
-            .sha512_ctx
-            .lock()
-            .map_err(|_e| InternalError::Unknown)?;
-        *guard = Some(ctx);
-        Ok(())
-    }
-
-    fn sha512_digest_update(&self, data: &[u8]) -> Result<(), InternalError> {
-        let mut guard = self
-            .sha512_ctx
-            .lock()
-            .map_err(|_e| InternalError::Unknown)?;
-        if let Some(ref mut ctx) = *guard {
-            ctx.update(data).map_err(|_e| InternalError::Unknown)?;
-        }
-        Ok(())
-    }
-
-    fn sha512_digest_final(&self) -> Result<Vec<u8>, InternalError> {
-        let mut guard = self
-            .sha512_ctx
-            .lock()
-            .map_err(|_e| InternalError::Unknown)?;
-        if let Some(ref mut ctx) = *guard {
-            ctx.finish()
-                .map(|buf| buf.as_ref().to_owned())
-                .map_err(|_e| InternalError::Unknown)
-        } else {
-            Err(InternalError::Unknown)
-        }
-    }
-
-    fn encrypt(
-        &self,
-        cipher: SignalCipherType,
-        key: &[u8],
-        iv: &[u8],
-        data: &[u8],
-    ) -> Result<Vec<u8>, InternalError> {
-        use openssl::symm::Mode;
-        self.crypter(Mode::Encrypt, cipher, key, iv, data)
-    }
-
-    fn decrypt(
-        &self,
-        cipher: SignalCipherType,
-        key: &[u8],
-        iv: &[u8],
-        data: &[u8],
-    ) -> Result<Vec<u8>, InternalError> {
-        use openssl::symm::Mode;
-        self.crypter(Mode::Decrypt, cipher, key, iv, data)
-    }
 }
 
 /// A simple vtable ([`signal_crypto_provider`]) and set of trampolines to let C
